feat: show active periods in TimeInfo components
previously, we only showed current periods
